% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BookingRetrieval.R
\name{BookingRetrieval}
\alias{BookingRetrieval}
\title{Orchestrates retrieving the bookings from iLab}
\usage{
BookingRetrieval(
  data,
  chromote_session,
  theurl = "https://cibr.umaryland.edu/schedules",
  FolderName = "ScheduleMonitor",
  AlternateDirectory = NULL,
  PauseInterval = 5
)
}
\arguments{
\item{data}{data.frame with "TheInstrument" column, containing the
instrument names, followed by a "Code" column with the instrument url hash}

\item{chromote_session}{The returned chromote session from BrowserRun()}

\item{theurl}{Default is "https://cibr.umaryland.edu/schedules", swap 
in your own institutions iLab page url}

\item{FolderName}{Default ScheduleMonitor, can switch 
out as desired}

\item{AlternateDirectory}{Default is your Documents folder,
can switch out if desired}

\item{PauseInterval}{Seconds between headless browser steps to allow page to load.}
}
\value{
CSV files to the data folder, and a data.frame to R.
}
\description{
Orchestrates retrieving the bookings from iLab
}
\examples{

A <- 2 + 2

}
