% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrowserRun.R
\name{BrowserRun}
\alias{BrowserRun}
\title{This Function Returns a Running Chromote Session}
\usage{
BrowserRun(
  theurl = "https://cibr.umaryland.edu/schedules",
  instrumenthash = "512453#",
  FolderName = "ScheduleMonitor",
  AlternateDirectory = NULL,
  pattern = NULL,
  PauseInterval = 5
)
}
\arguments{
\item{theurl}{Default is "https://cibr.umaryland.edu/schedules", swap 
in your own institutions iLab page url}

\item{instrumenthash}{Default is "512453#", swap 
in your own instruments url hash}

\item{FolderName}{Default ScheduleMonitor, can switch 
out as desired}

\item{AlternateDirectory}{Default is your Documents folder,
can switch out if desired}

\item{pattern}{For # usethis::edit_r_environ(), use THE_USER and THE_PASS}

\item{PauseInterval}{Seconds between headless browser steps to allow page to load.}
}
\value{
An active chromote session post login
}
\description{
This Function Returns a Running Chromote Session
}
\examples{

A <- 2 + 2


}
