% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FolderSetup.R
\name{FolderSetup}
\alias{FolderSetup}
\title{Creates a "ScheduleMonitor" folder under Documents.}
\usage{
FolderSetup(FolderName = "ScheduleMonitor", AlternateDirectory = NULL)
}
\arguments{
\item{FolderName}{Default is ScheduleMonitor, can switch 
out as desired}

\item{AlternateDirectory}{Default is your Documents folder,
can switch out if desired}
}
\value{
A ScheduleMonitor folder with its subfolders.
}
\description{
Creates a "ScheduleMonitor" folder under Documents.
}
\examples{

A <- 2 + 2

}
