% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iLabPlots.R
\name{iLabPlots}
\alias{iLabPlots}
\title{Generates the schedule plots in an iLab format}
\usage{
iLabPlots(
  x,
  data,
  FolderName = "ScheduleMonitor",
  AlternateDirectory = NULL,
  width = 8.5,
  height = 11,
  units = "in",
  dpi = 300,
  device = "png"
)
}
\arguments{
\item{x}{The name of the instrument}

\item{data}{The returned parsed booking data from BookingRetrieval}

\item{FolderName}{Default ScheduleMonitor, can switch 
out as desired}

\item{AlternateDirectory}{Default is your Documents folder,
can switch out if desired}

\item{width}{plot default 8.5}

\item{height}{plot default 11}

\item{units}{plot default "in"}

\item{dpi}{plot default 300}

\item{device}{"png"}
}
\value{
A plot for display
}
\description{
Generates the schedule plots in an iLab format
}
\examples{

A <- 2 + 2

}
