% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PasswordSet.R
\name{KeyringCheck}
\alias{KeyringCheck}
\title{Check whether keyring is correctly set up.}
\usage{
KeyringCheck(KeyringIgnore = FALSE)
}
\arguments{
\item{KeyringIgnore}{Ignores the keyring check, proceeds with the
use of the env variable}
}
\value{
TRUE if keyring is setup or bypassed, otherwise false
}
\description{
Should work for Windows/MacOS, Linux can be configured correctly, 
but may require manual intervention. Provides heads-up before
proceeding to default env option.
}
\keyword{internal}
