% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PasswordSet.R
\name{PasswordSet}
\alias{PasswordSet}
\title{Uses the keyring package to set your password credentials}
\usage{
PasswordSet(KeyringIgnore = FALSE)
}
\arguments{
\item{KeyringIgnore}{Default FALSE, set to TRUE to proceed with
storing credentials in default env when secure keychain setup
is still incomplete (can occur on some Linux OS)}
}
\value{
Stores password for future use.
}
\description{
Uses the keyring package to set your password credentials
}
\examples{

A <- 2 + 2

}
