% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaspberryPatch.R
\name{RaspberryPatch}
\alias{RaspberryPatch}
\title{Buils the .R and .sh scripts needed to automate}
\usage{
RaspberryPatch(FolderName = "ScheduleMonitor", AlternateDirectory = NULL)
}
\arguments{
\item{FolderName}{Default is ScheduleMonitor, can switch 
out as desired}

\item{AlternateDirectory}{Default is your Documents folder,
can switch out if desired}
}
\value{
.R and .sh scripts to the FolderName folder
}
\description{
Buils the .R and .sh scripts needed to automate
}
\examples{

A <- 2+2

}
