% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iLabPlots.R
\name{iLabPlots}
\alias{iLabPlots}
\title{Generates the schedule plots in an iLab format}
\usage{
iLabPlots(x, data)
}
\arguments{
\item{x}{TBD}

\item{data}{TBD}
}
\value{
A plot for display
}
\description{
Generates the schedule plots in an iLab format
}
\examples{

A <- 2 + 2

}
