% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iLabSummaryPlot.R
\name{iLabSummaryPlot}
\alias{iLabSummaryPlot}
\title{Returns a instrument summary plot for the particular day}
\usage{
iLabSummaryPlot(data)
}
\arguments{
\item{data}{TBD}

\item{x}{TBD}

\item{instruments}{TBD}
}
\value{
A summary plot
}
\description{
Returns a instrument summary plot for the particular day
}
\examples{

A <- 2+2

}
