% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PasswordSet.R
\name{PasswordSet}
\alias{PasswordSet}
\title{Uses the keyring package to set your password credentials}
\usage{
PasswordSet(KeyringIgnore = FALSE)
}
\arguments{
\item{KeyringIgnore}{If TRUE, proceeds with storing credentials
in environment variables when no secure keychain is available. 
Default is FALSE.}
}
\value{
Securely stores password on your local system
}
\description{
Uses the keyring package to set your password credentials
}
\examples{

A <- 2 + 2

}
