% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PasswordSet.R
\name{UsernameSet}
\alias{UsernameSet}
\title{Uses the keyring package to set your username credentials}
\usage{
UsernameSet(KeyringIgnore = FALSE)
}
\arguments{
\item{KeyringIgnore}{If TRUE, proceeds with storing credentials
in environment variables when no secure keychain is available. 
Default is FALSE.}
}
\value{
Securely stores username on your local system
}
\description{
Uses the keyring package to set your username credentials
}
\examples{

A <- 2 + 2

}
